test: consolidate download size-limit tests into one module
The wget and curl size-limit tests were near-identical copies in their
respective test files, tripping SonarCloud's duplicated-lines gate.
Run both commands through one parameterized test case instead.
