+++ src/cowrie/test/test_wget.py
@@ -17,12 +17,10 @@ from urllib.parse import urlparse
 
 from twisted.internet import defer, error
 from twisted.python.failure import Failure
-from twisted.web.http_headers import Headers
 
 from cowrie.commands.wget import Command_wget
 from cowrie.core.artifact import Artifact
 from cowrie.core.config import CowrieConfig
-from cowrie.core.network import DownloadLimitExceeded
 from cowrie.shell.command import HoneyPotCommand
 from cowrie.shell.protocol import HoneyPotInteractiveProtocol
 from cowrie.test.eventcapture import capture_events
@@ -185,91 +183,6 @@ class WgetArtifactCleanupTests(unittest.TestCase):
             [],
             "late download callbacks reported events for an exited command",
         )
-
-
-class FakeBodyTransport:
-    """Records whether the response body producer was stopped."""
-
-    def __init__(self) -> None:
-        self.stopped = False
-
-    def stopProducing(self) -> None:
-        self.stopped = True
-
-
-class FakeResponse:
-    """Minimal treq response: headers, length, and a recordable body."""
-
-    def __init__(self, length: int) -> None:
-        self.length = length
-        self.code = 200
-        self.phrase = b"OK"
-        self.headers = Headers({b"content-type": [b"application/octet-stream"]})
-        self.transport = FakeBodyTransport()
-        self.delivered_to: Any = None
-
-    def deliverBody(self, protocol: Any) -> None:
-        self.delivered_to = protocol
-        protocol.makeConnection(self.transport)
-
-
-class WgetSizeLimitTests(unittest.TestCase):
-    """Exceeding download_limit_size must abort the transfer, not merely stop
-    saving it (issue #1500)."""
-
-    def setUp(self) -> None:
-        self.proto = HoneyPotInteractiveProtocol(FakeAvatar(FakeServer()))
-        self.tr = FakeTransport("", "31337")
-        self.proto.makeConnection(self.tr)
-        self.tr.clear()
-
-        self.tmpdir = tempfile.mkdtemp()
-        self._orig_artifact_dir = Artifact.artifactDir
-        Artifact.artifactDir = self.tmpdir
-
-    def tearDown(self) -> None:
-        Artifact.artifactDir = self._orig_artifact_dir
-        self.proto.connectionLost()
-        for name in os.listdir(self.tmpdir):
-            os.remove(os.path.join(self.tmpdir, name))
-        os.rmdir(self.tmpdir)
-
-    def _make_cmd(self) -> Command_wget:
-        cmd = Command_wget.__new__(Command_wget)
-        cmd.protocol = self.proto
-        cmd.writefn = lambda _data: None
-        cmd.errorWritefn = lambda _data: None
-        cmd.exit_code = 0
-        cmd.url = b"http://198.51.100.1/big.bin"
-        cmd.host = "198.51.100.1"
-        cmd.port = 80
-        cmd.quiet = True
-        cmd.limit_size = 10
-        cmd.started = time.time()
-        cmd.artifact = Artifact("wget-download")
-        self.proto.cmdstack.append(cmd)
-        return cmd
-
-    # treq closes the connection when the collector raises, so collect must
-    # raise once the transfer exceeds the limit instead of draining the rest
-    # of the body into the void.
-    def test_over_limit_chunk_raises_to_close_connection(self) -> None:
-        cmd = self._make_cmd()
-        cmd.currentlength = 0
-        with self.assertRaises(DownloadLimitExceeded):
-            cmd.collect(b"x" * 20)
-        self.assertTrue(cmd.exited)
-
-    # A Content-Length over the limit must abort the body instead of leaving
-    # it undelivered (twisted buffers an undelivered body in memory while the
-    # server keeps sending).
-    def test_content_length_over_limit_aborts_body(self) -> None:
-        cmd = self._make_cmd()
-        response = FakeResponse(length=1000)
-        cmd.success(response)
-        self.assertTrue(cmd.exited)
-        self.assertIsNotNone(response.delivered_to)
-        self.assertTrue(response.transport.stopped)
 
 
 class WgetOutboundBindTests(unittest.TestCase):
